Stream Z: 29 P2 items + audit-5 CTCLoss follow-up doc
audit-5 P2.  Lands every Stream Z item.

Backend / nn / optim (Z.4–Z.13):
* Z.4 vulkan_ops_misc.cpp roi_align_backward gains BF16 narrow-back
  (F16 already had it).
* Z.5 oneapi_backend.cpp:637 warp_size selection now picks
  largest <= 64 from sub_group_sizes (was front()) so Intel iGPUs
  (16) / AMD-via-Codeplay (64) / NVIDIA SYCL (32) all get the right
  reduction tile size.
* Z.6 verified: Y.10 already added ensure_atomic_float_supported
  gate on ROIAlignBackward.
* Z.7 deleted dead fused_adam_step_{f16,bf16}.comp shaders;
  vulkan_ops_fused dispatchFusedAdamStep throws on F16/BF16 inputs
  (optim/adam.cpp:97 only dispatches fused on CUDA).
* Z.8 sparse_adam.cpp adds `if (param.has_sparse_grad()) continue;`
  after the fast-path so empty-nnz sparse grads don't fall through
  to dense-rows scan.
* Z.9 adam_atan2.cpp drops static_cast<float> casts on lr/beta/eps
  AttrKey writes; native double passed to fused CUDA path.
* Z.10 LRScheduler base gains make_scheduler_type_tensor /
  check_scheduler_type helpers in scheduler.hpp / scheduler.cpp.
  Wired state["scheduler_type"] write + check on load across all
  16 scheduler subclasses (Step, Exponential, CosineAnnealing,
  LinearWarmup, Constant, Linear, Multiplicative [force flag],
  Sequential, Chained, ReduceLROnPlateau, Cyclic, OneCycle,
  CosineAnnealingWarmRestarts, Lambda [force flag], MultiStep,
  Polynomial).  Mismatched type throws unless force=true.
* Z.11 lazy_linear.cpp forward picks requested_device_.value_or
  (input.device()); WARN_ONCE on mismatch.
* Z.12 nadam.cpp / adamax.cpp / lion.cpp gain state["num_params"]
  write + load guard (LAMB already had it).
* Z.13 fsdp.cpp gate widened to dtype_size(comm_dtype) <
  dtype_size(dtype); cast-back uses dtype (was hard-coded F32);
  WARN_ONCE when comm_dtype is no narrower than param dtype.

Python (Z.14–Z.22):
* Z.14 nn.pyi: dropped spurious padding_mode + Tuple[int,int,int]
  from Conv3d / ConvTranspose1d stubs (ConvTranspose3d already
  matched).
* Z.15 data.py: _worker_info is now thread-local;
  _PrefetchWorker (num_workers=1 thread fast path) installs
  WorkerInfo(id=0, num_workers=1, dataset) + calls worker_init_fn(0)
  before iteration, matching the multi-worker path.
* Z.16 DataLoader.__init__ raises ValueError if prefetch_factor < 1.
* Z.17 bindings_nn.cpp hook bindings wrap via WeakMethod (bound
  methods) / weakref.ref (other callables); skip callback if
  weakref expired.  Reference cycle broken for the common case.
* Z.18 bindings_nn.cpp get_submodules: py::return_value_policy::
  reference_internal so mutations propagate.
* Z.19 numpy_interop.cpp BFloat16 without ml_dtypes now raises
  ValueError with install instructions (was UserWarning + lossy
  uint16 roundtrip).
* Z.20 new autograd::pow(Variable, Variable) and pow(double,
  Variable) overloads implemented as exp(b * log(a)) — preserves
  grad_fn on both inputs.  Variable.__pow__(Variable, Variable)
  and __rpow__(Variable, float) wired in bindings_core.
* Z.21 new C++ free functions tenzor::nn::functional::
  {max,avg,adaptive_max,adaptive_avg}_pool{1,3}d (8 total)
  delegate to layer Module forward_impl; matching pybind bindings;
  Python wrappers in functional.py rewritten to call them directly
  (no more throwaway layer modules per call).  count_include_pad
  =False explicitly rejected (not supported by the layer).
* Z.22 Module.load_state_dict returns py::none() on strict=True
  success; (missing, unexpected) tuple only on strict=False.
  PyTorch _IncompatibleKeys contract restored.

Tests / tooling (Z.23–Z.29):
* Z.23 test_jit_trace.py leak check is now `assert leak == set(...)`
  with Pow + Tanh as forbidden decomposition primitives.
* Z.24 test_phase11_backends.cpp 8 raw GTEST_SKIPs migrated to
  SKIP_WITH_REASON(SkipReason::BackendUnavailable, …).
* Z.25 11 baseline entries migrated to EXPECT_GRAD_FLOWS:
  4× pooling1d (MaxPool/AvgPool/AdaptiveMaxPool/AdaptiveAvgPool
  GradientFlow), 4× pooling3d (same family), 3× rnn_cells_dispatch
  (LSTMCell, GRUCell, LSTM backward).
* Z.26 verified __array__ copy=True forces fresh allocation
  (W.16 path was correct); added 4 test cases to
  tests/python/test_array_protocol.py.
* Z.27 new tests/autograd/test_chunk_split.cpp (5 tests) covers:
  chunk(zeros({0,5}), 3, 0).size() == 3 with each output [0,5];
  split(zeros({0,5}), 2, 0).size() == 1 (single empty); plus
  destructured-binding regression for Y.5 + non-empty grad-flow.
  Two impl fixes required to make this pass:
  - autograd::split now returns a 1-element vector with empty slice
    when dim_size==0 (PyTorch parity).
  - Tensor::slice bounds check skips empty-result tensors (numel==0)
    — was spuriously throwing on [0:0] of an already-empty dim.
* Z.28 new test coverage for new audit-4/5 entry points:
  - new tests/autograd/test_save_backward_multi_device.cpp
    (custom Function TwoInputSaver saves tensors on dev0+dev1;
    skips when only one GPU visible).
  - new tests/python/test_variable_getitem.py (9 tests) covering
    int / negative-int / slice / strided slice / tuple / fancy /
    boolean-mask indexing, asserting var.grad flows after each.
  - U.4 / W.3 coverage already landed in Y.30 / Y.31.
* Z.29 11 additional EXPECT_NO_THROW patterns migrated in
  test_losses_advanced_multidtype.cpp (KLDiv, Focal, Dice, Huber,
  BCEWithLogits, CTC, TripletMarginWithDistance, MultiLabelMargin,
  InfoNCE, NTXent, Triplet).  Baseline reduced from 34 (post-Z.25)
  to 23 — 32% reduction beyond Z.25.

Discovered follow-up (docs/known-bugs/audit5-ctcloss-runtime-
correctness.md):
The Y.31 fix unmasked a real audit-5 follow-up — CUDA/ROCm/OneAPI
CTCLoss kernels (added in W.3) produce +inf on T=50 inputs (missing
log-sum-exp max-shift); Vulkan has no kernel.  Recorded as a
multi-day audit-6-scope item, with required next steps and CPU
reference pointer.

Verified:
* ninja build clean (single -j1 shell).
* tools/lint_test_contract.py exits 0 (868 test files, 23
  pre-existing violations all in baseline, 0 new).
* tools/parity_coverage_matrix.py exits 0 from build/.
* All targeted ctest selectors run clean.

+++ src/nn/optim/adam_atan2.cpp
@@ -94,11 +94,13 @@ auto AdamAtan2::step_impl() -> void {
             }
 
             NewOpAttributes attrs;
-            attrs.set(AttrKey::Lr, static_cast<float>(hp.lr));
-            attrs.set(AttrKey::Beta1, static_cast<float>(hp.beta1));
-            attrs.set(AttrKey::Beta2, static_cast<float>(hp.beta2));
-            attrs.set(AttrKey::Eps, static_cast<float>(hp.eps));
-            attrs.set(AttrKey::WeightDecay, static_cast<float>(hp.weight_decay));
+            // Z.9: pass native double to match the fused-Adam contract; the
+            // previous static_cast<float> narrowed Float64 params' hyperparams.
+            attrs.set(AttrKey::Lr, hp.lr);
+            attrs.set(AttrKey::Beta1, hp.beta1);
+            attrs.set(AttrKey::Beta2, hp.beta2);
+            attrs.set(AttrKey::Eps, hp.eps);
+            attrs.set(AttrKey::WeightDecay, hp.weight_decay);
             attrs.set(AttrKey::Step, step_count_);
             attrs.set(AttrKey::Amsgrad, amsgrad_);
